Refactor feature 1 naming
