+++ .github/workflows/test.yml
@@ -72,7 +72,7 @@ jobs:
     runs-on: ubuntu-latest
     needs: build
     env:
-      FEATURES: "F1"
+      FEATURES: "feature_1"
     steps:
       - uses: actions/checkout@v4.2.2
       - name: Start dev container
