+++ packages/home_index/main.py
@@ -87,7 +87,7 @@ PROJECT_ROOT = Path(__file__).resolve().parents[2]
 if str(PROJECT_ROOT) not in sys.path:
     sys.path.insert(0, str(PROJECT_ROOT))
 
-from features.F1 import scheduler as F1_scheduler
+from features.feature_1 import scheduler as F1_scheduler
 
 
 # endregion
